Fix popping up autocomplete if input not selected
If the user typed in the autocomplete input but subsequently blurred
the field before results came back from the server (by, say, tabbing
to the next form element), the asynchronous return of results would
end up rendering the suggestion list anyway. This is jarring and
unexpected behavior. This commit fixes this bug.

Change-Id: I3534e8ef12ff3aab565f332f29b8c01043ffde1c
Reviewed-on: https://gerrit.causes.com/19108
Reviewed-by: Alex Wallace <alex.wallace@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Tony Wooster <tony.wooster@causes.com>