Add `onActivate` callback
This commit adds an `onActivate` callback, which is triggered when the
user activates -- via mouse or keyboard -- an item in the autocomplete
list. This allows a little bit more customization for autocomplete
fields that behave more like inline-search where replacing the value
of the input element is undesirable.

This commit also fixes a bad interaction between `isSelectable` and
`autoSelectFirst`, whereby the first item in the list would be
selected regardless of its `isSelectable` status.

Change-Id: Ia7886314cb6827bce04bb94cd325f706e707bdc7
Reviewed-on: https://gerrit.causes.com/19293
Reviewed-by: Tony Wooster <tony.wooster@causes.com>
Tested-by: Tony Wooster <tony.wooster@causes.com>