Fix single-character deletion bug
Because the setting of `currentValue` on input change was delayed a
minute amount, a return from the server of no selectable suggestions
could cause the last character in the search input to be deleted. This
commit fixes this bug by immediately updating `currentValue`.

Additionally, this commit sets the value from the canonical `el.val()`
rather than from `element.value` which brings the plugin in line with
itself.

Change-Id: Ifdb0638314aa647fd6151ee49c61025eaae4de3e
Reviewed-on: https://gerrit.causes.com/19368
Reviewed-by: Tony Wooster <tony.wooster@causes.com>
Tested-by: Tony Wooster <tony.wooster@causes.com>