Add customizable bad-query check
The autocomplete's bad query check assumes that any query that begins
with a previous query that returned no results should also be a bad
query. E.g., if the user types "foo" and gets no results, "food"
should noop and not hit the server. This commit adds the ability to
customize this method such that an exact match or other handling can
be subbed instead.

Change-Id: I3f724bb66a547b3a1629dd74a6382466e9661b44
Reviewed-on: https://gerrit.causes.com/20437
Reviewed-by: Aiden Scandella <aiden@causes.com>
Tested-by: Tony Wooster <tony.wooster@causes.com>